backend/requirements.txt
fastapi
uvicorn
pillow
stepic
python-multipart
boto3
python-dotenv
pydub
numpy